api/requirements.txt
# Web Framework & Server
fastapi==0.128.0
gunicorn==23.0.0
uvicorn==0.40.0

# Validation & Configuration
pydantic==2.12.5
python-dotenv==1.2.1
python-decouple==3.8

# Machine Learning
scikit-learn==1.4.2
joblib==1.3.2
numpy==1.26.4
pandas==2.1.4

# LangChain & AI
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langchain-ollama==0.1.0
langchain-openai==0.1.22
toonkit==0.1.1

# Utilities
dependency-injector==4.48.3
urllib3==2.6.3
httpx==0.27.0